Fixed small bug in cuda implementation
git-svn-id: http://www.mpa-garching.mpg.de/svn/cosmo-group/kdolag/splotch@23033 019ef673-9300-0410-806d-f16013bd1951

+++ cuda/splotch.cu
@@ -303,7 +303,7 @@ long int cu_get_chunk_particle_count(CuPolicy* policy, size_t psize, int ntiles,
    size_t spareMem = 20*(1<<20);
    long int arrayParticleSize = gMemSize - 4*ImSize - 2*tiles - spareMem;
    long int len = (long int) (arrayParticleSize/((psize+2*sizeof(int))*pfactor)); 
-   long int maxlen = policy->GetMaxGridSize() * policy->GetBlockSize();
+   long int maxlen = (long int)policy->GetMaxGridSize() * (long int)policy->GetBlockSize();
    if (len > maxlen) len = maxlen;
    return len;
   }
